Split up WIRC code nicely
